+++ crypto/winapi/CryptoAlgoritmProvider.h
@@ -15,7 +15,7 @@ namespace CommonLib
 				static	CBcryptAlgProviderPtr Create(LPCWSTR pszAlgId, ULONG dwFlags);
 				~CBcryptAlgProvider();
 
-				void CryptDeriveKeyPBKDF2(const crypto_astr& password, const crypto_vector& salt, uint64_t interations, uint32_t keySize, crypto_vector& derivedKey);
+				void CryptDeriveKeyPBKDF2(const crypto_astr& password, const crypto_vector& salt, uint32_t interations, uint32_t keySize, crypto_vector& derivedKey);
 				void CryptGenRandom(byte* pBuf, uint32_t size, ULONG dwFlags);
 				BCRYPT_KEY_HANDLE CryptGenerateSymmetricKey(PUCHAR pbKeyObject, ULONG cbKeyObject, PUCHAR  pbSecret, ULONG  cbSecret, ULONG dwFlags);
 				void CryptGetProperty(LPCWSTR pszProperty, PUCHAR  pbOutput, ULONG  cbOutput, ULONG  *pcbResult, ULONG dwFlags);
